requirements.txt
mysqlclient==2.2.4
python-dotenv==1.0.1